refactor: 오타수정 및 메서드 네이밍 통일
