+++ src/main/java/com/example/falleterbev2/domain/feed/presentation/dto/response/FeedAllResponse.java
@@ -1,4 +1,4 @@
-package com.example.falleterbev2.domain.feed.presentation.dto.reponse;
+package com.example.falleterbev2.domain.feed.presentation.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
